requirements.txt
awscli==1.32.74
boto3==1.34.64
click==8.1.7
datasets==2.18.0
huggingface_hub==0.22.1
pandas==2.2.1
Pillow==10.2.0
python-dotenv==1.0.1
s3fs # versioning this breaks the install
scikit_learn==1.4.1.post1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2